Move time_series aggregation to aggregations module.
This also drops the TimeSeries ceremonial interface.

Note that this change also moves a search cancellation test to aggregations module. It does so by creating a base search cancallation base class that both server and this module share.

Relates to #90283
Relates to #82273
