--- server/src/test/java/org/elasticsearch/search/aggregations/timeseries/TimeSeriesAggregationBuilderTests.java
@@ -1,59 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0 and the Server Side Public License, v 1; you may not use this file except
- * in compliance with, at your election, the Elastic License 2.0 or the Server
- * Side Public License, v 1.
- */
-
-package org.elasticsearch.search.aggregations.timeseries;
-
-import org.elasticsearch.common.io.stream.NamedWriteableRegistry;
-import org.elasticsearch.common.io.stream.Writeable;
-import org.elasticsearch.common.settings.Settings;
-import org.elasticsearch.search.SearchModule;
-import org.elasticsearch.search.aggregations.AggregatorFactories;
-import org.elasticsearch.test.AbstractXContentSerializingTestCase;
-import org.elasticsearch.xcontent.NamedXContentRegistry;
-import org.elasticsearch.xcontent.XContentParser;
-
-import java.io.IOException;
-import java.util.Collections;
-
-import static org.hamcrest.Matchers.hasSize;
-
-public class TimeSeriesAggregationBuilderTests extends AbstractXContentSerializingTestCase<TimeSeriesAggregationBuilder> {
-
-    @Override
-    protected TimeSeriesAggregationBuilder doParseInstance(XContentParser parser) throws IOException {
-        assertSame(XContentParser.Token.START_OBJECT, parser.nextToken());
-        AggregatorFactories.Builder parsed = AggregatorFactories.parseAggregators(parser);
-        assertThat(parsed.getAggregatorFactories(), hasSize(1));
-        assertThat(parsed.getPipelineAggregatorFactories(), hasSize(0));
-        TimeSeriesAggregationBuilder agg = (TimeSeriesAggregationBuilder) parsed.getAggregatorFactories().iterator().next();
-        assertNull(parser.nextToken());
-        assertNotNull(agg);
-        return agg;
-    }
-
-    @Override
-    protected Writeable.Reader<TimeSeriesAggregationBuilder> instanceReader() {
-        return TimeSeriesAggregationBuilder::new;
-    }
-
-    @Override
-    protected TimeSeriesAggregationBuilder createTestInstance() {
-        return new TimeSeriesAggregationBuilder(randomAlphaOfLength(10), randomBoolean());
-    }
-
-    @Override
-    protected NamedWriteableRegistry getNamedWriteableRegistry() {
-        return new NamedWriteableRegistry(new SearchModule(Settings.EMPTY, Collections.emptyList()).getNamedWriteables());
-    }
-
-    @Override
-    protected NamedXContentRegistry xContentRegistry() {
-        return new NamedXContentRegistry(new SearchModule(Settings.EMPTY, Collections.emptyList()).getNamedXContents());
-    }
-
-}
